Feat/code preview

+++ packages/cli/src/commands/add.ts
@@ -11,6 +11,8 @@ import { logger } from "../utils/logger";
 import { custom, z } from "zod";
 import { getPackageManager } from "../utils/get-package-manager";
 import { framework_supports } from "../utils/get-suppoted";
+import { type Config } from "tailwindcss/types/config";
+import { TAILWIND_CONFIG } from "../utils/templates";
 
 process.on("SIGINT", () => process.exit(0));
 process.on("SIGTERM", () => process.exit(0));
@@ -24,7 +26,7 @@ const addCommandInput = z.object({
   id: z.string(),
   cwd: z.string(),
 });
-
+const TAILWIND_EXTEND_PROPS = ["animation", "keyframes"];
 console.log(gradient("pink", "blue")(FARMUI_GRAFFITI));
 
 export const add = new Command()
@@ -131,6 +133,57 @@ export const add = new Command()
       const path_to_add: CompToAddProps[] = [];
       // for now , the content we will support will be react based , toll we have updated the ednpoint
       const root_comp_name = select_files_by_id.files[0].root.name;
+      const tailwind_values = select_files_by_id["tailwind"];
+      if (tailwind_values) {
+        logger.info("The components need tailwind.config.(ts/js) change");
+
+        const tailwind_config_js = path.join(
+          process.cwd(),
+          "tailwind.config.js"
+        );
+        const tailwind_config_ts = path.join(
+          process.cwd(),
+          "tailwind.config.ts"
+        );
+        let config_exists = "ts";
+        let is_config_exist = true;
+        if (
+          !existsSync(tailwind_config_ts) &&
+          !existsSync(tailwind_config_js)
+        ) {
+          is_config_exist = false;
+        }
+        if (!is_config_exist) {
+          logger.warn(
+            "The tailwind.config.(js/ts) can not be found with in the given directory , Please make sure to intialize your project with tailwindcss or shadcn , may be any UI component that build on top of tailwind"
+          );
+          logger.info("Skipping... and importing the components");
+        } else {
+          // means the either the js or tailwind config exists
+          if (existsSync(tailwind_config_js)) {
+            config_exists = "js";
+          }
+          const tailwind_config_file = `tailwind.config.${config_exists}`;
+
+          const tailwin_config_path = path.join(
+            process.cwd(),
+            tailwind_config_file
+          );
+          const read_tailwind_config = await fs.readFile(
+            tailwin_config_path,
+            "utf8"
+          );
+          TAILWIND_EXTEND_PROPS.map((props) => {
+            if (props in tailwind_values) {
+              logger.info(
+                `For ${props} - copy the following and paste it to your ${tailwind_config_file}`
+              );
+              const tailwind_props_value = tailwind_values[props].content;
+              console.log({ [props]: tailwind_props_value });
+            }
+          });
+        }
+      }
       const root_comp_content =
         select_files_by_id.files[0].root.contents[framework].content;
       const root_comp_path = path.join(root_dir, root_comp_name);
@@ -154,13 +207,35 @@ export const add = new Command()
         });
       });
       const spinner = ora(`Dumping your components...`);
-      spinner.start();
+      let spinner_overwrite = ora("Overriting");
       const dependencies: string[] = select_files_by_id.dependencies;
       if (!path_to_add) {
         logger.warn("No component to add");
       } else {
         path_to_add.map(async ({ comp_content, comp_path }) => {
-          await fs.writeFile(`${comp_path}.tsx`, comp_content);
+          const comp_write_path = `${comp_path}.tsx`;
+          if (existsSync(comp_write_path)) {
+            const { overwrite } = await prompts({
+              type: "confirm",
+              name: "overwrite",
+              message: `Override already existed component? Proceed?`,
+              initial: true,
+            });
+            if (overwrite) {
+              spinner_overwrite = ora(
+                `Overwriting the component at ${comp_write_path}....`
+              );
+              spinner_overwrite.start();
+              await fs.writeFile(`${comp_path}.tsx`, comp_content);
+            } else {
+              logger.info(
+                `Skipping: The component ${comp_write_path} already existed`
+              );
+            }
+          } else {
+            spinner.start();
+            await fs.writeFile(`${comp_path}.tsx`, comp_content);
+          }
         });
       }
       const packageManager = await getPackageManager(custom_cwd);
@@ -170,6 +245,7 @@ export const add = new Command()
         });
       }
       spinner.stop();
+      spinner_overwrite.stop();
       if (dependencies.length) {
         logger.info(`Dependencies - ${dependencies.length} added`);
         dependencies.map((dep) => {
